Separate kernel tuners in .cpp with main and .hpp with settings
